ia-agent-tickets-v2/agent/agent-v2/pyproject.toml
[project]
name = "agent-v2"
version = "0.1.0"
description = "Pluggable AI agent for enterprise helpdesk ticket systems (Odoo-ready)"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.28",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "uvicorn[standard]>=0.44.0",
]
